+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Metadata
   * Updated MSCloudLoginAssistant to 1.0.15;
+  * Updated SharePointPnPPowerShellOnline to 3.21.2005.1;
 
 ## 1.0.5.128